refactor(system): move client to system
