+++ app/react/portainer/system/build-url.ts
